feat(rpc): SPI & serialization, now serialization and transport can be configured

+++ rpc-common/src/main/java/cn/krone/rpc/common/exchange/RpcRequest.java
@@ -28,7 +28,7 @@ public class RpcRequest extends RpcMessage { // 因为要网络传输实现 Seri
     private Object[] args;
 
     @Override
-    public int getMessageType() {
+    public byte getMessageType() {
         return RPC_MESSAGE_TYPE_REQUEST;
     }
 }
